CAUSEWAY-3897: reinstate contract tests (RO)
