+++ viewers/restfulobjects/pom.xml
@@ -40,7 +40,6 @@
         <module>viewer</module>
         <module>test</module>
         <module>testing</module>
-<!--        <module>jaxrs-resteasy</module>-->
     </modules>
 
 </project>
