+++ viewers/restfulobjects/rendering/pom.xml
@@ -17,8 +17,10 @@
   specific language governing permissions and limitations
   under the License.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <groupId>org.apache.causeway.viewer</groupId>
@@ -28,11 +30,13 @@
 
     <artifactId>causeway-viewer-restfulobjects-rendering</artifactId>
 
-	<name>Apache Causeway Viewer - RO (Rendering)</name>
+    <name>Apache Causeway Viewer - RO (Rendering)</name>
 
     <properties>
-        <jar-plugin.automaticModuleName>org.apache.causeway.viewer.restfulobjects.rendering</jar-plugin.automaticModuleName>
-        <git-plugin.propertiesDir>org/apache/causeway/viewer/restfulobjects/rendering</git-plugin.propertiesDir>
+        <jar-plugin.automaticModuleName>
+            org.apache.causeway.viewer.restfulobjects.rendering</jar-plugin.automaticModuleName>
+        <git-plugin.propertiesDir>
+            org/apache/causeway/viewer/restfulobjects/rendering</git-plugin.propertiesDir>
     </properties>
 
 
@@ -45,7 +49,13 @@
             <groupId>org.apache.causeway.viewer</groupId>
             <artifactId>causeway-viewer-restfulobjects-applib</artifactId>
         </dependency>
-        
+
+        <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+        </dependency>
+
+
         <!-- TEST DEPENDENCIES -->
         <dependency>
             <groupId>org.apache.causeway.core</groupId>
@@ -53,7 +63,7 @@
             <scope>test</scope>
         </dependency>
 
-	</dependencies>
+    </dependencies>
 
 
 </project>
