+++ viewers/restfulobjects/rendering/src/main/java/org/apache/causeway/viewer/restfulobjects/rendering/context/SerializationStrategy.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.viewer.resources.serialization;
+package org.apache.causeway.viewer.restfulobjects.rendering.context;
 
 import java.util.Collection;
