+++ viewers/restfulobjects/rendering/src/main/java/org/apache/causeway/viewer/restfulobjects/rendering/exhandling/ExceptionDetail.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.viewer.exhandling.entity;
+package org.apache.causeway.viewer.restfulobjects.rendering.exhandling;
 
 import java.util.List;
 import java.util.stream.Stream;
