+++ viewers/restfulobjects/rendering/src/test/java/org/apache/causeway/viewer/restfulobjects/rendering/context/ResourceContext_stripQuotes_Test.java
@@ -16,13 +16,15 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.viewer.context;
+package org.apache.causeway.viewer.restfulobjects.rendering.context;
 
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
+
+import org.apache.causeway.viewer.restfulobjects.rendering.context.ResourceContext;
 
 class ResourceContext_stripQuotes_Test {
