+++ viewers/restfulobjects/testing/pom.xml
@@ -45,7 +45,7 @@
 
         <dependency>
             <groupId>org.apache.causeway.viewer</groupId>
-            <artifactId>causeway-viewer-restfulobjects-viewer</artifactId>
+            <artifactId>causeway-viewer-restfulobjects-rendering</artifactId>
         </dependency>
 
         <dependency>
