+++ viewers/restfulobjects/viewer/src/main/java/org/apache/causeway/viewer/restfulobjects/viewer/CausewayModuleViewerRestfulObjectsViewer.java
@@ -29,9 +29,9 @@ import org.apache.causeway.applib.services.exceprecog.RootCauseFinder;
 import org.apache.causeway.core.webapp.CausewayModuleCoreWebapp;
 import org.apache.causeway.viewer.commons.services.CausewayModuleViewerCommonsServices;
 import org.apache.causeway.viewer.restfulobjects.rendering.CausewayModuleRestfulObjectsRendering;
+import org.apache.causeway.viewer.restfulobjects.rendering.exhandling.ExceptionResponseFactory;
 import org.apache.causeway.viewer.restfulobjects.rendering.service.acceptheader.AcceptHeaderServiceForRest;
 import org.apache.causeway.viewer.restfulobjects.viewer.exhandling.ExceptionMapperForRestfulObjects;
-import org.apache.causeway.viewer.restfulobjects.viewer.exhandling.ExceptionResponseFactory;
 import org.apache.causeway.viewer.restfulobjects.viewer.resources.DomainObjectResourceServerside;
 import org.apache.causeway.viewer.restfulobjects.viewer.resources.DomainServiceResourceServerside;
 import org.apache.causeway.viewer.restfulobjects.viewer.resources.DomainTypeResourceServerside;
