+++ viewers/restfulobjects/viewer/src/test/java/org/apache/causeway/viewer/restfulobjects/viewer/RepresentationTypeTest_getMediaType.java
@@ -16,10 +16,11 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.jaxrsresteasy;
+package org.apache.causeway.viewer.restfulobjects.viewer;
 
-import org.apache.causeway.viewer.restfulobjects.testing.Parser_forCacheControl_ContractTest;
+import org.apache.causeway.viewer.restfulobjects.testing.RepresentationTypeTest_getMediaType_ContractTest;
 
-public class Parser_forCacheControl_Test extends Parser_forCacheControl_ContractTest {
+public class RepresentationTypeTest_getMediaType 
+extends RepresentationTypeTest_getMediaType_ContractTest {
 
 }
