+++ viewers/restfulobjects/viewer/src/test/java/org/apache/causeway/viewer/restfulobjects/viewer/RepresentationTypeTest_getName.java
@@ -16,30 +16,10 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.viewer.resources;
+package org.apache.causeway.viewer.restfulobjects.viewer;
 
-import org.apache.causeway.applib.annotation.Where;
-import org.apache.causeway.viewer.restfulobjects.applib.RepresentationType;
-import org.apache.causeway.viewer.restfulobjects.rendering.service.RepresentationService;
+import org.apache.causeway.viewer.restfulobjects.testing.RepresentationTypeTest_getName_ContractTest;
 
-/**
- * @since 2.0
- */
-public record ResourceDescriptor(
-        RepresentationType representationType,
-        Where where,
-        RepresentationService.Intent intent,
-        ResourceLink resourceLink) {
-
-    public enum ResourceLink {
-        NONE,
-        OBJECT,
-        SERVICE
-    }
-
-    public static ResourceDescriptor empty() {
-        // in support of testing
-        return new ResourceDescriptor(null, null, null, ResourceLink.NONE);
-    }
+public class RepresentationTypeTest_getName extends RepresentationTypeTest_getName_ContractTest {
 
 }
