+++ viewers/restfulobjects/viewer/src/test/java/org/apache/causeway/viewer/restfulobjects/viewer/RepresentationTypeTest_lookup.java
@@ -16,10 +16,10 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.jaxrsresteasy;
+package org.apache.causeway.viewer.restfulobjects.viewer;
 
-import org.apache.causeway.viewer.restfulobjects.testing.Parser_forJaxRsMediaType_ContractTest;
+import org.apache.causeway.viewer.restfulobjects.testing.RepresentationTypeTest_lookup_ContractTest;
 
-public class Parser_forJaxRsMediaType_Test extends Parser_forJaxRsMediaType_ContractTest {
+public class RepresentationTypeTest_lookup extends RepresentationTypeTest_lookup_ContractTest {
 
 }
