+++ viewers/restfulobjects/viewer/src/test/java/org/apache/causeway/viewer/restfulobjects/viewer/RepresentationTypeTest_parser.java
@@ -16,10 +16,10 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.restfulobjects.jaxrsresteasy;
+package org.apache.causeway.viewer.restfulobjects.viewer;
 
-import org.apache.causeway.viewer.restfulobjects.testing.RepresentationTypeTest_lookup_ContractTest;
+import org.apache.causeway.viewer.restfulobjects.testing.RepresentationTypeTest_parser_ContractTest;
 
-public class RepresentationTypeTest_lookup extends RepresentationTypeTest_lookup_ContractTest {
+public abstract class RepresentationTypeTest_parser extends RepresentationTypeTest_parser_ContractTest {
 
 }
